Delay the check for node
My node and npm commands are loaded on demand, via some lazy binding. I
do this to speed up the shell launch time. There's probably a good way
to do this via the new zplugin infrastructure.

I dropped the nvm version from the prompt as it was too wide and
unnecessary. Knowing the ambient node verison is more important to me.