Remove INFO and TITLE
Hitting <ENTER> with no command has a weird side-effect with geometry:
it shows the INFO like background jobs and host-name. There wasn't an
easy way to turn this off, so I'm removing it for now.

Also changing the title bar for the terminal: not sure how useful this
either. In the interest of minimizing the surface area of what geometry
does by default, removing this too.